Not sure if this is a good spot for py2app patches, if not please let me know.
This fixes an issue with copying a different version of Python over to an app/plugin bundle than the one used to run py2app with.

Before this patch you could run into situations where if you run py2app under Python 2.6, and it tried to copy the Python 2.5 framework over to the app bundle, it would be at some point try to copy a path like:

'/Library/Frameworks/Python.framework/Versions/2.5/include/python2.6/pyconfig.h'

Because it used sys.version instead of the version number in the framework info passed in, to figure out what the last folder in the path should be, this simple patch fixes it.

Patch by Dirk Stoop.