Two tweaks for issue #132
1) Depend on modulegraph 0.12

2) Explicitly report modules that cannot be found and
   modules with syntax errors.

With these changes it is possible to use py2app in projects
that contain separate modules with python 2 or python 3
compatibility code with syntax that is not valid in the
other major release.

The report for missing modules is still a bit rough, modulegraph
falsely reports missing modules for "from collections import defaultdict".