Report about invalid relative imports
With this commit py2app will report about modules that contain
invalid relative imports, that is relative imports either outside
of a package, or refering to a path that moves outside of a toplevel
package.

This also fixes a problem where py2app would crap out when encountering
such an import.