Include package metadata in the generated bundle
The information from relevant ".dist-info" and ".egg-info"
directories is now included in the generated application bundle
(in the zip file). This does *not* include the record of installed
files.

The code supports both the modern "RECORD" file and older
"installed-files.txt" because I noticed both in a fresh python 3.9
virtual env.

Also add a recipe for gcloud, autoinclude "google.apis" due to
a dependency in package metadata not detected by the dependency scanner.

Fixes #1, fixes #328