Workaround for setuptools distribution autodiscovery
This introduces a ``setuptools.finalize_distribution_options``
that sets the *name* and *py_modules* attributes of the
distribution in a way that is compatible with py2app's regular
calculation of these attributes.

This avoid problems with setuptools 61.0.0 or later that
use autodiscovery to set these attributes to values that may
not be compatible with py2app.

Issue #414