Add an "maybe_packages" option
This option behaves like "packages", but only
if the parts of the named packages are detected
by the dependency checker.

The new feature is used by the recipes as well,
this avoids force inclusion of a package when
a recipe for a node that's not reachable from the
root includes that package.