remove dependencies on pkg_resources where possible
Use importlib.resources and importlib.metadata
instead of pkg_resources.  The only code left
using pkg_resources is in support of that library.

Also enforce this using the flake8 pre-commit check.

This also adds a PySide6 example and fixes the PySide
recipe. Note that both are very minimal and are primarily
used for testing.