Next bunch of updates
This is a commit after a long pause, hence the vague
commit message.

The code is now building a complete app bundle, but
that bundle is still not complete.

Next up:
- Use the new launcher code in the bundle
- Generate the __boot__.py file

That should get us close to generating a working app
bundle using the new code paths (for app bundles,
plugin bundles need more work).

Note that the code isn't tested beyond manually running
the new builder on a small sample project, proper testing
will start when I'm happy with the overall design.