Implement 'python3' command in the app bundle
Before this commit the 'python3' command was a copy of the
normal stub, it now supports the same interface as python itself
but runs in isolated mode, does not the user site directory and
cannot be copied outside of the bundle.

SomeApp.app/Contents/MacOS/python3 is basically a python3 command
in the app's context.